third_party/blink/web_tests/external/wpt/editing/run/multitest_3001-4000-expected.txt

This is a testharness.js-based test.
Found 198 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["superscript",""],["justifycenter",""]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["justifycenter",""]] "foo[]bar" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""],["justifycenter",""]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["justifycenter",""]] "foo[]bar" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["superscript",""],["justifycenter",""]] "foo[]bar" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["superscript",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\">foo<sup>a</sup>bar</div>" but got "<div style=\\"text-align:center\\">fooabar</div>"
[FAIL] [["superscript",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["superscript",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["superscript",""],["justifyfull",""]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["justifyfull",""]] "foo[]bar" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""],["justifyfull",""]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["justifyfull",""]] "foo[]bar" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["superscript",""],["justifyfull",""]] "foo[]bar" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["superscript",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\">foo<sup>a</sup>bar</div>" but got "<div style=\\"text-align:justify\\">fooabar</div>"
[FAIL] [["superscript",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["superscript",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["superscript",""],["justifyleft",""]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["justifyleft",""]] "foo[]bar" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""],["justifyleft",""]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["justifyleft",""]] "foo[]bar" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["superscript",""],["justifyleft",""]] "foo[]bar" queryCommandValue("justifyleft") after
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["superscript",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sup>a</sup>bar" but got "fooabar"
[FAIL] [["superscript",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["superscript",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyleft") after
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["superscript",""],["justifyright",""]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["justifyright",""]] "foo[]bar" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""],["justifyright",""]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["justifyright",""]] "foo[]bar" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["superscript",""],["justifyright",""]] "foo[]bar" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["superscript",""],["justifyright",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\">foo<sup>a</sup>bar</div>" but got "<div style=\\"text-align:right\\">fooabar</div>"
[FAIL] [["superscript",""],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["superscript",""],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["superscript",""],["outdent",""]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["outdent",""]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["superscript",""],["outdent",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sup>a</sup>bar" but got "foo<span style=\\"vertical-align:super\\">a</span>bar"
[FAIL] [["superscript",""],["outdent",""],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["outdent",""],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<u>a</u>bar" but got "foo<span style=\\"text-decoration-line:underline\\">a</span>bar"
[FAIL] [["underline",""],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""],["delete",""]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["delete",""]] "foo[]bar" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["underline",""],["delete",""]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["delete",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "fo<u>a</u>bar" but got "foabar"
[FAIL] [["underline",""],["delete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["delete",""],["inserttext","a"]] "foo[]bar" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["underline",""],["delete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["formatblock","<div>"]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["formatblock","<div>"]] "foo[]bar" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["underline",""],["formatblock","<div>"]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo<u>a</u>bar</div>" but got "<div>fooabar</div>"
[FAIL] [["underline",""],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["underline",""],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["forwarddelete",""]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["forwarddelete",""]] "foo[]bar" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["underline",""],["forwarddelete",""]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["forwarddelete",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<u>a</u>ar" but got "fooaar"
[FAIL] [["underline",""],["forwarddelete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["forwarddelete",""],["inserttext","a"]] "foo[]bar" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["underline",""],["forwarddelete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["indent",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote>foobar</blockquote>" but got "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">foobar</blockquote>"
[FAIL] [["underline",""],["indent",""]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["indent",""]] "foo[]bar" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["underline",""],["indent",""]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["indent",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote>foo<u>a</u>bar</blockquote>" but got "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">fooabar</blockquote>"
[FAIL] [["underline",""],["indent",""],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["indent",""],["inserttext","a"]] "foo[]bar" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["underline",""],["indent",""],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["inserthorizontalrule",""]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["inserthorizontalrule",""]] "foo[]bar" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["underline",""],["inserthorizontalrule",""]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["inserthorizontalrule",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<hr><u>a</u>bar" but got "foo<hr>abar"
[FAIL] [["underline",""],["inserthorizontalrule",""],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["inserthorizontalrule",""],["inserttext","a"]] "foo[]bar" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["underline",""],["inserthorizontalrule",""],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["inserthtml","ab<b>c</b>d"]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["inserthtml","ab<b>c</b>d"]] "foo[]bar" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["underline",""],["inserthtml","ab<b>c</b>d"]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["inserthtml","ab<b>c</b>d"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "fooab<b>c</b>d<u>a</u>bar" but got "fooab<b>c</b>dabar"
[FAIL] [["underline",""],["inserthtml","ab<b>c</b>d"],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["inserthtml","ab<b>c</b>d"],["inserttext","a"]] "foo[]bar" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["underline",""],["inserthtml","ab<b>c</b>d"],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["insertimage","/img/lion.svg"]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["insertimage","/img/lion.svg"]] "foo[]bar" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["underline",""],["insertimage","/img/lion.svg"]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["insertimage","/img/lion.svg"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<img src=\\"/img/lion.svg\\"><u>a</u>bar" but got "foo<img src=\\"/img/lion.svg\\">abar"
[FAIL] [["underline",""],["insertimage","/img/lion.svg"],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["insertimage","/img/lion.svg"],["inserttext","a"]] "foo[]bar" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["underline",""],["insertimage","/img/lion.svg"],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["insertlinebreak",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br>bar" but got "foo<span style=\\"text-decoration-line:underline\\"><br></span>bar"
[FAIL] [["underline",""],["insertlinebreak",""]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["insertlinebreak",""]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""],["insertlinebreak",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br><u>a</u>bar" but got "foo<span style=\\"text-decoration-line:underline\\"><br>a</span>bar"
[FAIL] [["underline",""],["insertlinebreak",""],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["insertlinebreak",""],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""],["insertorderedlist",""]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["insertorderedlist",""]] "foo[]bar" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["underline",""],["insertorderedlist",""]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["insertorderedlist",""]] "foo[]bar" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["insertorderedlist",""]] "foo[]bar" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo<u>a</u>bar</li></ol>" but got "<ol><li>afoobar</li></ol>"
[FAIL] [["underline",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["underline",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""],["insertparagraph",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p>" but got "foo<div>bar</div>"
[FAIL] [["underline",""],["insertparagraph",""]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["insertparagraph",""]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""],["insertparagraph",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p><u>a</u>bar</p>" but got "foo<div><span style=\\"text-decoration-line:underline\\">a</span>bar</div>"
[FAIL] [["underline",""],["insertparagraph",""],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["insertparagraph",""],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["insertunorderedlist",""]] "foo[]bar" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["underline",""],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo<u>a</u>bar</li></ul>" but got "<ul><li>afoobar</li></ul>"
[FAIL] [["underline",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["underline",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""],["justifycenter",""]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["justifycenter",""]] "foo[]bar" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["underline",""],["justifycenter",""]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["justifycenter",""]] "foo[]bar" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["underline",""],["justifycenter",""]] "foo[]bar" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["underline",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\">foo<u>a</u>bar</div>" but got "<div style=\\"text-align:center\\">fooabar</div>"
[FAIL] [["underline",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["underline",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["underline",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["underline",""],["justifyfull",""]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["justifyfull",""]] "foo[]bar" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["underline",""],["justifyfull",""]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["justifyfull",""]] "foo[]bar" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["underline",""],["justifyfull",""]] "foo[]bar" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["underline",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\">foo<u>a</u>bar</div>" but got "<div style=\\"text-align:justify\\">fooabar</div>"
[FAIL] [["underline",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["underline",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["underline",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["underline",""],["justifyleft",""]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["justifyleft",""]] "foo[]bar" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["underline",""],["justifyleft",""]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["justifyleft",""]] "foo[]bar" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["underline",""],["justifyleft",""]] "foo[]bar" queryCommandValue("justifyleft") after
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["underline",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<u>a</u>bar" but got "fooabar"
[FAIL] [["underline",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["underline",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["underline",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyleft") after
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["underline",""],["justifyright",""]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["justifyright",""]] "foo[]bar" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["underline",""],["justifyright",""]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["justifyright",""]] "foo[]bar" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["underline",""],["justifyright",""]] "foo[]bar" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["underline",""],["justifyright",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\">foo<u>a</u>bar</div>" but got "<div style=\\"text-align:right\\">fooabar</div>"
[FAIL] [["underline",""],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["underline",""],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["underline",""],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["underline",""],["outdent",""]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["outdent",""]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""],["outdent",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<u>a</u>bar" but got "foo<span style=\\"text-decoration-line:underline\\">a</span>bar"
[FAIL] [["underline",""],["outdent",""],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""],["outdent",""],["inserttext","a"]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["backcolor","#00FFFF"],["delete",""]] "foo[]bar" queryCommandValue("backcolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 255, 255)" but got "rgba(0, 0, 0, 0)"
[FAIL] [["backcolor","#00FFFF"],["delete",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "fo<span style=\\"background-color:rgb(0, 255, 255)\\">a</span>bar" but got "foabar"
[FAIL] [["backcolor","#00FFFF"],["delete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("backcolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 255, 255)" but got "rgba(0, 0, 0, 0)"
[FAIL] [["backcolor","#00FFFF"],["formatblock","<div>"]] "foo[]bar" queryCommandValue("backcolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 255, 255)" but got "rgba(0, 0, 0, 0)"
[FAIL] [["backcolor","#00FFFF"],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo<span style=\\"background-color:rgb(0, 255, 255)\\">a</span>bar</div>" but got "<div>fooabar</div>"
[FAIL] [["backcolor","#00FFFF"],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" queryCommandValue("backcolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 255, 255)" but got "rgba(0, 0, 0, 0)"
[FAIL] [["backcolor","#00FFFF"],["forwarddelete",""]] "foo[]bar" queryCommandValue("backcolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 255, 255)" but got "rgba(0, 0, 0, 0)"
[FAIL] [["backcolor","#00FFFF"],["forwarddelete",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"background-color:rgb(0, 255, 255)\\">a</span>ar" but got "fooaar"
[FAIL] [["backcolor","#00FFFF"],["forwarddelete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("backcolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 255, 255)" but got "rgba(0, 0, 0, 0)"
[FAIL] [["backcolor","#00FFFF"],["indent",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote>foobar</blockquote>" but got "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">foobar</blockquote>"
[FAIL] [["backcolor","#00FFFF"],["indent",""]] "foo[]bar" queryCommandValue("backcolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 255, 255)" but got "rgba(0, 0, 0, 0)"
[FAIL] [["backcolor","#00FFFF"],["indent",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote>foo<span style=\\"background-color:rgb(0, 255, 255)\\">a</span>bar</blockquote>" but got "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">fooabar</blockquote>"
[FAIL] [["backcolor","#00FFFF"],["indent",""],["inserttext","a"]] "foo[]bar" queryCommandValue("backcolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 255, 255)" but got "rgba(0, 0, 0, 0)"
Harness: the test ran to completion.
